refactor : 함수이름 수정
