+++ src/main/java/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package calcul;
+package calculator;
 
 public class Calculator {
     public int add(int a, int b) {
