+++ src/main/java/racingcar/Winner.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
-public class NonTestRacingCar {
+public class Winner {
 
     RacingCar racingCar = new RacingCar();
