+++ src/test/java/calculatortest/CalculTest.java
@@ -1,6 +1,6 @@
-package calcultest;
+package calculatortest;
 
-import calcul.Calculator;
+import calculator.Calculator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
